Move device IDs from settings to backend
